Refactor de la estructura de tests files
